clean up pathfora tests
