+++ test/acceptance/gate.spec.js
@@ -1,4 +1,9 @@
 import globalReset from '../utils/global-reset';
+import {
+  createSiteGateWidget,
+  expectWidgetVisible,
+  expectWidgetClosed
+} from '../utils/test-helpers';
 
 // -------------------------
 //  GATE
@@ -9,7 +14,7 @@ describe('the gate component', function () {
   });
 
   it('should open gate when the record is not set', function (done) {
-    var gate = new pathfora.SiteGate({
+    var gate = createSiteGateWidget({
       headline: 'Blocking Widget',
       id: 'sitegate-widget-1',
       msg: 'Submit this widget to access the website.'
@@ -17,16 +22,14 @@ describe('the gate component', function () {
 
     pathfora.initializeWidgets([gate]);
 
-    var widget = $('#' + gate.id);
-
     setTimeout(function () {
-      expect(widget.hasClass('opened')).toBeTruthy();
+      expectWidgetVisible(gate.id);
       done();
     }, 200);
   });
 
   it('should should work with showForm: false', function (done) {
-    var gate = new pathfora.SiteGate({
+    var gate = createSiteGateWidget({
       id: 'gate-hide-form',
       headline: 'Gated Site Feature',
       msg: 'Please agree to the terms to proceed.',
@@ -38,16 +41,14 @@ describe('the gate component', function () {
       pathfora.initializeWidgets([gate]);
     }).not.toThrow();
 
-    var widget = $('#' + gate.id);
-
     setTimeout(function () {
-      expect(widget.hasClass('opened')).toBeTruthy();
+      expectWidgetVisible(gate.id);
       done();
     }, 200);
   });
 
   it('should not gate when the record is already set', function (done) {
-    var gate = new pathfora.SiteGate({
+    var gate = createSiteGateWidget({
       headline: 'Blocking Widget',
       id: 'sitegate-widget-2',
       msg: 'Submit this widget to access the website.'
@@ -57,10 +58,8 @@ describe('the gate component', function () {
 
     pathfora.initializeWidgets([gate]);
 
-    var widget = $('#' + gate.id);
-
     setTimeout(function () {
-      expect(widget.hasClass('opened')).toBeFalsy();
+      expectWidgetClosed(gate.id);
       done();
     }, 200);
   });
