+++ test/acceptance/scaffolding.spec.js
@@ -1,4 +1,5 @@
 import globalReset from '../utils/global-reset';
+import { createMessageWidget, createFormWidget } from '../utils/test-helpers';
 
 // -------------------------
 // SCAFFOLDING
@@ -18,7 +19,7 @@ describe('when building a scaffolding component', function () {
   it('should insert widget into config after building and inserting into scaffold', function () {
     var scaffold = pathfora.utils.initWidgetScaffold();
 
-    var tester = new pathfora.Message({
+    var tester = createMessageWidget({
       id: 'tester123',
       headline: 'Sample Insert',
       msg: 'Sample insert message.',
@@ -45,7 +46,7 @@ describe('when building a scaffolding component', function () {
   it('should insert multiple widgets into config binding to the same segment', function () {
     var scaffold = pathfora.utils.initWidgetScaffold();
 
-    var tester1 = new pathfora.Message({
+    var tester1 = createMessageWidget({
       id: 'tester123',
       headline: 'Sample Insert',
       msg: 'Sample insert message.',
@@ -56,7 +57,7 @@ describe('when building a scaffolding component', function () {
     });
     pathfora.utils.insertWidget('target', 'smt_new', tester1, scaffold);
 
-    var tester2 = new pathfora.Form({
+    var tester2 = createFormWidget({
       id: 'tester456',
       headline: 'Sample Insert Two',
       msg: 'Sample insert message two.',
@@ -85,7 +86,7 @@ describe('when building a scaffolding component', function () {
   it('should insert multiple widgets into config binding to the same segment but excluding', function () {
     var scaffold = pathfora.utils.initWidgetScaffold();
 
-    var tester1 = new pathfora.Message({
+    var tester1 = createMessageWidget({
       id: 'tester123',
       headline: 'Sample Insert',
       msg: 'Sample insert message.',
@@ -97,7 +98,7 @@ describe('when building a scaffolding component', function () {
 
     pathfora.utils.insertWidget('exclude', 'smt_new', tester1, scaffold);
 
-    var tester2 = new pathfora.Form({
+    var tester2 = createFormWidget({
       id: 'tester456',
       headline: 'Sample Insert Two',
       msg: 'Sample insert message two.',
